+++ src/token.ts
@@ -1,3 +1,5 @@
-export function getTokenPath(): string {
-  return `.obsidian/todoist-token`;
+import type { Vault } from "obsidian";
+
+export function getTokenPath(vault: Vault): string {
+  return `${vault.configDir}/todoist-token`;
 }