+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - You can now press 'Enter' in the create task dialog to create the task. [Thanks to @oleeskild](https://github.com/oleeskild)
 
+### 🐛 Bug Fixes
+
+- The plugin should now respect your Vault configuration directory when saving your Todoist token.
+
 ### [1.9.0] - 2022-01-19
 
 ### ✨ Features